pyproject.toml
[project]
name = "steyeartsite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.3",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.0",
]
